requirements.txt
openai==0.28.1
google-generativeai==0.4.1
pinecone-client==3.0.0
python-dotenv==1.0.0
